# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Gestión Académica",
        "description": "API REST para la gestión de cursos, estudiantes, docentes y materias",
        "contact": {
            "name": "Soporte API",
            "email": "admin@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor Local"
        }
    ],
    "paths": {
...
        "/api/materia-curso": {
            "get": {
                "tags": [
                    "Asignaciones Materia-Curso"
                ],
                "summary": "Obtener todas las asignaciones de materias a cursos",
                "operationId": "4e1d550e6033dcc506386c20218618ca",
                "responses": {
                    "200": {
                        "description": "Lista de asignaciones obtenida correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Asignaciones obtenidas correctamente"
                                        },
                                        "errors": {
                                            "type": "null"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "materias_id_materias": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "cursos_id_cursos": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "materia": {
                                                        "properties": {
                                                            "id_materia": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "nombre": {
                                                                "type": "string",
                                                                "example": "Matemáticas"
                                                            },
                                                            "estado": {
                                                                "type": "string",
                                                                "example": "A"
                                                            }
                                                        },
                                                        "type": "object"
                                                    },
                                                    "curso": {
                                                        "properties": {
                                                            "id_curso": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "nombre": {
                                                                "type": "string",
                                                                "example": "Matemáticas Básicas"
                                                            },
                                                            "estado": {
                                                                "type": "string",
                                                                "example": "A"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No se encontraron asignaciones"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Asignaciones Materia-Curso"
                ],
                "summary": "Crear una nueva asignación de materia a curso",
                "operationId": "1ee0071440dbc69bf6ce578f8e21d140",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "materias_id_materias",
                                    "cursos_id_cursos"
                                ],
                                "properties": {
                                    "materias_id_materias": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "cursos_id_cursos": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Asignación creada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 201
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Asignación creada correctamente"
                                        },
                                        "errors": {
                                            "type": "null"
                                        },
                                        "data": {
                                            "properties": {
                                                "materia_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "curso_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error de validación"
                    },
                    "409": {
                        "description": "La materia ya está asignada a este curso"
                    },
                    "500": {
                        "description": "Error del servidor"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}